requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
autoflake==2.3.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black @ git+https://github.com/psf/black.git@6fdf8a4af28071ed1d079c01122b34c5d587207a
blinker==1.7.0
Brotli==1.1.0
build==1.1.1
cachetools==5.3.3
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.19.0
crewai-tools==0.0.15
cryptography==42.0.5
cssselect2==0.7.0
dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
docstring-parser==0.15
embedchain==0.1.93
emoji==2.10.1
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
griffe==0.41.3
groq==0.4.2
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
identify==2.5.35
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
iniconfig==2.0.0
instructor==0.5.2
isort==5.13.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-google-genai==0.0.11
langchain-groq==0.0.1
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.19
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==1.0.1
mkdocs-material==9.5.12
mkdocs-material-extensions==1.3.1
mkdocstrings==0.22.0
mkdocstrings-python==1.8.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
paginate==0.5.6
pandas==2.2.1
pathspec==0.12.1
peewee==3.17.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
posthog==3.5.0
pre-commit==3.6.2
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
py==1.11.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pyflakes==3.2.0
PyGithub==1.59.1
Pygments==2.17.2
PyJWT==2.8.0
pylance==0.9.18
pymdown-extensions==10.7.1
PyNaCl==1.5.0
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.352
pysbd==0.3.4
PySocks==1.7.1
pytest==8.0.2
pytest-vcr==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
pytils==0.4.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
rapidfuzz==3.6.1
ratelimiter==1.2.0.post0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
schema==0.7.5
sec-api==1.0.17
selenium==4.18.1
semver==3.0.2
serpapi==0.1.5
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.32.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tools==0.1.9
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.21.0
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
vcrpy==6.0.1
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.37
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
